feat: modified checkbox
